Bullets Refactor

* Testing

* Merge branch 'develop' into feature/bullets-refactor

* Tags

* Fix Speed

* Cleanup

* Merge branch 'develop' into feature/bullets-refactor

* Improvements

* Rebase

* Test

* Remove dlls


Approved-by: Miha Krajnc
Approved-by: Nik Denissov

+++ Quantum/quantum_code/quantum.code/Configs/Constants.cs
@@ -16,7 +16,7 @@ namespace Quantum
 		public static readonly int BURST_INTERVAL_DIVIDER = 3;
 		public static readonly FP SELF_DAMAGE_MODIFIER= FP._0_75;
 		public static readonly FP PICKUP_SPEED_MINIMUM = FP._0_10;
-		public static readonly int[] APPRX_NORMAL_DISTRIBUTION = {1, 13, 35, 65, 87, 99, 100};
+		public static readonly double[] APPRX_NORMAL_DISTRIBUTION = {1, 13, 35, 65, 87, 99, 100};
 		public static readonly FP TAP_TO_USE_SPECIAL_AIMING_OFFSET = FP._0_75 + FP._0_10;
 		public static readonly FP LOW_AMMO_THRESHOLD_TO_DROP_MORE = FP._0_50 + FP._0_10;
